chore(appveyor): Edit appveyor and remove old example
