--- examples/microphone.py
@@ -1,105 +1,0 @@
-# You need to install pyaudio to run this example
-# pip install pyaudio
-
-from __future__ import print_function
-import pyaudio
-from watson_developer_cloud import SpeechToTextV1
-from watson_developer_cloud.websocket import RecognizeCallback, AudioSource
-from threading import Thread
-
-try:
-    from Queue import Queue, Full
-except ImportError:
-    from queue import Queue, Full
-
-# initialize speech to text service
-speech_to_text = SpeechToTextV1(
-    username='YOUR SERVICE USERNAME',
-    password='YOUR SERVICE PASSWORD',
-    url='https://stream.watsonplatform.net/speech-to-text/api')
-
-# define callback for the service
-class MyRecognizeCallback(RecognizeCallback):
-    def __init__(self):
-        RecognizeCallback.__init__(self)
-
-    def on_transcription(self, transcript):
-        print(transcript)
-
-    def on_connected(self):
-        print('Connection was successful')
-
-    def on_error(self, error):
-        print('Error received: {}'.format(error))
-
-    def on_inactivity_timeout(self, error):
-        print('Inactivity timeout: {}'.format(error))
-
-    def on_listening(self):
-        print('Service is listening')
-
-    def on_hypothesis(self, hypothesis):
-        print(hypothesis)
-
-    def on_data(self, data):
-        print(data)
-
-    def on_close(self):
-        print("Connection closed")
-
-# Variables for recording the speech
-FORMAT = pyaudio.paInt16
-CHANNELS = 1
-RATE = 44100
-CHUNK = 1024
-# Note: It will discard if the websocket client can't consumme fast enough
-BUF_MAX_SIZE = CHUNK * 10
-
-# Buffer to store audio
-q = Queue(maxsize=int(round(BUF_MAX_SIZE / CHUNK)))
-
-# define callback to store the recording in queue
-def callback(in_data, frame_count, time_info, status):
-    try:
-        q.put(in_data)
-    except Full:
-        pass # discard
-    return (None, pyaudio.paContinue)
-
-# get ready with service params
-audio_source = AudioSource(q, True, True)
-def recognize_using_weboscket(*args):
-    mycallback = MyRecognizeCallback()
-    speech_to_text.recognize_using_websocket(audio=audio_source,
-                                             content_type='audio/l16; rate=44100',
-                                             recognize_callback=mycallback)
-
-# instantiate pyaudio
-audio = pyaudio.PyAudio()
-
-# open stream using callback
-stream = audio.open(
-    format=FORMAT,
-    channels=CHANNELS,
-    rate=RATE,
-    input=True,
-    frames_per_buffer=CHUNK,
-    stream_callback=callback,
-    start=False
-)
-
-print("Enter CTRL+C to end recording...")
-stream.start_stream()
-
-try:
-    recognize_thread = Thread(target=recognize_using_weboscket, args=())
-    recognize_thread.start()
-
-    while True:
-        pass
-except KeyboardInterrupt:
-    # stop recording
-    audio_source.completed_recording()
-    stream.stop_stream()
-    stream.close()
-    audio.terminate()
